document server struct
